requirements.txt
numpy
tensorflow
mtcnn
flask
pillow
pip install -r requirements.txt
